move rosetta models to rosetta package
