+++ api/rosetta/retriever.go
@@ -15,12 +15,12 @@
 package rosetta
 
 import (
-	"github.com/optakt/flow-dps/models/identifier"
-	"github.com/optakt/flow-dps/models/rosetta"
+	"github.com/optakt/flow-dps/rosetta/identifier"
+	"github.com/optakt/flow-dps/rosetta/object"
 )
 
 type Retriever interface {
-	Block(network identifier.Network, block identifier.Block) (*rosetta.Block, []identifier.Transaction, error)
-	Transaction(network identifier.Network, block identifier.Block, transaction identifier.Transaction) (*rosetta.Transaction, error)
-	Balances(network identifier.Network, block identifier.Block, account identifier.Account, currencies []identifier.Currency) ([]rosetta.Amount, error)
+	Block(network identifier.Network, block identifier.Block) (*object.Block, []identifier.Transaction, error)
+	Transaction(network identifier.Network, block identifier.Block, transaction identifier.Transaction) (*object.Transaction, error)
+	Balances(network identifier.Network, block identifier.Block, account identifier.Account, currencies []identifier.Currency) ([]object.Amount, error)
 }
