+++ rosetta/identifier/block.go
