+++ rosetta/identifier/currency.go
