+++ rosetta/identifier/operation.go
