+++ rosetta/object/error.go
@@ -12,7 +12,7 @@
 // License for the specific language governing permissions and limitations under
 // the License.
 
-package rosetta
+package object
 
 // Error is used to return rich errors from the API instead of utilizing HTTP
 // status codes (which often do not have a good analog). Both the code and the
